retail_store/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
amqp==1.4.9
anyjson==0.3.3
appconfig==0.1
appdirs==1.4.4
asgiref==3.4.1
async==0.6.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Baker==1.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.3.0.23
bs4==0.0.1
cchardet==2.1.7
celery==3.1.18
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
commander==0.2.3
cryptography==3.4.7
distlib==0.3.1
distro==1.5.0
Django==3.2.6
django-crispy-forms==1.13.0
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.20.1
drf-spectacular-sidecar==2021.9.27
filelock==3.0.12
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-SQLAlchemy==2.5.1
flatten-dict==0.4.1
greenlet==1.0.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
kombu==3.0.37
Markdown==3.3.4
MarkupSafe==2.0.1
model-bakery==1.3.2
more-itertools==8.8.0
multidict==5.1.0
numpy==1.21.1
packaging==20.9
paramiko==2.7.2
peewee==3.14.4
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycares==4.0.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.18.0
pytools==2021.2.7
pytz==2018.5
pywin32==227
PyYAML==5.4.1
requests==2.26.0
salary==2.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.4.23
sqlparse==0.4.1
sudo==1.0.0
texttable==1.6.4
toml==0.10.2
tqdm==4.56.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.6
vk-api==11.9.1
websocket-client==0.59.0
Werkzeug==2.0.1
yarl==1.6.3
